🌱 Split oversized core infra card components
Removes the `/* eslint-disable max-lines */` suppressions from four
oversized card components by extracting cohesive sibling modules.
Behaviour-preserving: all logic and markup moved verbatim.

KubeKong.tsx (830 -> 492)
  - KubeKong.constants.ts   game constants, geometry, types
  - KubeKong.draw.ts        canvas scene rendering
  - KubeKongOverlays.tsx    HUD + start/pause/game-over overlays

Checkers.tsx (848 -> 465)
  - Checkers.types.ts       board types, constants, taunt copy
  - Checkers.engine.ts      move generation, minimax AI, persistence
  - CheckersPiece.tsx       single-piece rendering

ClusterCosts.tsx (876 -> 522)
  - ClusterCosts.constants.ts     pricing tables, provider detection
  - ClusterCostsRatesPanel.tsx    pricing-rates info panel
  - ClusterCostsRow.tsx           per-cluster breakdown row
  - ClusterCostsFooter.tsx        attribution + methodology links

workload-monitor/LLMdStackMonitor.tsx (879 -> 653)
  - LLMdStackMonitor.constants.ts  types, sort/filter options, status maps
  - LLMdClusterFilter.tsx          cluster filter flyout
  - LLMdComponentSections.tsx      hierarchical component list
  - LLMdIssuesList.tsx             issues list

Fixes #21614
Fixes #21615
Refs #21503, #15790
