+++ web/src/components/cards/workload-monitor/LLMdStackMonitor.tsx
@@ -1,20 +1,17 @@
-/* eslint-disable max-lines -- TODO: split this file (tracked by #15790) */
 // Modal safety: the ApiKeyPromptModal imported here uses BaseModal with its own
 // close controls, and the cluster-filter dropdown is an anchored flyout (not a
 // backdrop modal). closeOnBackdropClick={false} semantics apply to the inline
 // inputs — no unsaved-changes risk from accidental backdrop clicks.
 import { useMemo, useState, useRef, useEffect } from 'react'
-import { createPortal } from 'react-dom'
 import {
   Cpu, Network, Activity, Layers, Server,
-  RefreshCw, ChevronDown, ChevronRight, Filter,
-  AlertTriangle
+  RefreshCw, AlertTriangle
 } from 'lucide-react'
 import { ALERT_SEVERITY_ORDER } from '../../../types/alerts'
 import { Skeleton } from '../../ui/Skeleton'
 import { Pagination } from '../../ui/Pagination'
 import { CardControls } from '../../ui/CardControls'
-import { CardSearchInput, CardAIActions } from '../../../lib/cards/CardComponents'
+import { CardSearchInput } from '../../../lib/cards/CardComponents'
 import { useCachedLLMdServers, useCachedGPUNodes } from '../../../hooks/useCachedData'
 import { useWorkloadMonitor } from '../../../hooks/useWorkloadMonitor'
 import { useDiagnoseRepairLoop } from '../../../hooks/useDiagnoseRepairLoop'
@@ -23,79 +20,29 @@ import { cn } from '../../../lib/cn'
 // WorkloadMonitorAlerts replaced with inline issue cards in Issues tab
 import { useLLMdClusters } from '../workload-detection/shared'
 import { useClusters } from '../../../hooks/useMCP'
-import { ClusterStatusDot, getClusterState } from '../../ui/ClusterStatusBadge'
-import { StatusBadge } from '../../ui/StatusBadge'
 import { useCardLoadingState } from '../CardDataContext'
 import type { MonitorIssue, MonitoredResource } from '../../../types/workloadMonitor'
 import { useTranslation } from 'react-i18next'
-
-type SortField = 'name' | 'status' | 'type' | 'cluster'
-type StatusFilter = 'all' | 'healthy' | 'degraded' | 'unhealthy'
-type IssueSortField = 'title' | 'severity' | 'cluster'
-type SeverityFilter = 'all' | 'critical' | 'warning' | 'info'
-
-const SORT_OPTIONS = [
-  { value: 'name', label: 'Name' },
-  { value: 'status', label: 'Status' },
-  { value: 'type', label: 'Type' },
-  { value: 'cluster', label: 'Cluster' },
-]
-
-const ISSUE_SORT_OPTIONS = [
-  { value: 'severity', label: 'Severity' },
-  { value: 'title', label: 'Title' },
-  { value: 'cluster', label: 'Cluster' },
-]
-
-const STATUS_FILTER_OPTIONS = [
-  { value: 'all', label: 'All' },
-  { value: 'healthy', label: 'Healthy' },
-  { value: 'degraded', label: 'Degraded' },
-  { value: 'unhealthy', label: 'Unhealthy' },
-]
-
-const SEVERITY_FILTER_OPTIONS = [
-  { value: 'all', label: 'All' },
-  { value: 'critical', label: 'Critical' },
-  { value: 'warning', label: 'Warning' },
-  { value: 'info', label: 'Info' },
-]
-
-const STATUS_ORDER: Record<string, number> = {
-  unhealthy: 0,
-  degraded: 1,
-  healthy: 2,
-  unknown: 3 }
+import { LLMdClusterFilter } from './LLMdClusterFilter'
+import { LLMdComponentSections, type LLMdSection } from './LLMdComponentSections'
+import { LLMdIssuesList } from './LLMdIssuesList'
+import {
+  ISSUE_SORT_OPTIONS,
+  SEVERITY_FILTER_OPTIONS,
+  SORT_OPTIONS,
+  STATUS_BADGE,
+  STATUS_FILTER_OPTIONS,
+  STATUS_ORDER,
+  type ComponentItem,
+  type IssueSortField,
+  type SeverityFilter,
+  type SortField,
+  type StatusFilter,
+} from './LLMdStackMonitor.constants'
 
 interface LLMdStackMonitorProps {
   config?: Record<string, unknown>
 }
-
-
-interface ComponentItem {
-  name: string
-  status: 'healthy' | 'degraded' | 'unhealthy' | 'unknown'
-  type?: string
-  namespace?: string
-  detail?: string
-  cluster?: string
-}
-
-const STATUS_DOT: Record<string, string> = {
-  healthy: 'bg-green-400',
-  degraded: 'bg-yellow-400',
-  unhealthy: 'bg-red-400',
-  unknown: 'bg-gray-400',
-  running: 'bg-green-400',
-  scaling: 'bg-yellow-400',
-  stopped: 'bg-red-400',
-  error: 'bg-red-400' }
-
-const STATUS_BADGE: Record<string, string> = {
-  healthy: 'bg-green-500/20 text-green-400',
-  degraded: 'bg-yellow-500/20 text-yellow-400',
-  unhealthy: 'bg-red-500/20 text-red-400',
-  unknown: 'bg-gray-500/20 dark:bg-gray-400/20 text-muted-foreground' }
 
 export function LLMdStackMonitor({ config: _config }: LLMdStackMonitorProps) {
   const { t } = useTranslation()
@@ -297,7 +244,7 @@ export function LLMdStackMonitor({ config: _config }: LLMdStackMonitorProps) {
   const needsPagination = itemsPerPage !== 'unlimited' && totalItems > limit
 
   // Build component sections from paginated items (for hierarchical view)
-  const sections = (() => {
+  const sections: LLMdSection[] = (() => {
     const SECTION_CONFIG: Array<{ type: string; label: string; icon: typeof Cpu; color: string }> = [
       { type: 'model', label: 'Model Serving', icon: Cpu, color: 'text-purple-400' },
       { type: 'epp', label: 'EPP', icon: Layers, color: 'text-blue-400' },
@@ -516,76 +463,17 @@ export function LLMdStackMonitor({ config: _config }: LLMdStackMonitorProps) {
         </span>
         {/* Cluster filter */}
         {availableClusters.length >= 1 && (
-          <div ref={clusterFilterRef} className="relative">
-            <button
-              ref={clusterFilterBtnRef}
-              onClick={() => setShowClusterFilter(!showClusterFilter)}
-              className={`flex items-center gap-1 px-2 py-1 text-xs rounded-lg border transition-colors ${
-                localClusterFilter.length > 0
-                  ? 'bg-purple-500/20 border-purple-500/30 text-purple-400'
-                  : 'bg-secondary border-border text-muted-foreground hover:text-foreground'
-              }`}
-              title="Filter by cluster"
-            >
-              <Filter className="w-3 h-3" />
-              {localClusterFilter.length > 0 && (
-                <span className="flex items-center gap-1">
-                  <Server className="w-3 h-3" />
-                  {localClusterFilter.length}/{availableClusters.length}
-                </span>
-              )}
-              <ChevronDown className="w-3 h-3" />
-            </button>
-            {showClusterFilter && dropdownStyle && createPortal(
-              <div
-                className="fixed w-48 max-h-48 overflow-y-auto rounded-lg bg-card border border-border shadow-lg z-50"
-                style={{ top: dropdownStyle.top, left: dropdownStyle.left }}
-                onMouseDown={e => e.stopPropagation()}
-              >
-                <div className="p-1">
-                  <button
-                    onClick={() => setLocalClusterFilter([])}
-                    className={`w-full px-2 py-1.5 text-xs text-left rounded transition-colors ${
-                      localClusterFilter.length === 0 ? 'bg-purple-500/20 text-purple-400' : 'hover:bg-secondary text-foreground'
-                    }`}
-                  >
-                    All clusters
-                  </button>
-                  {availableClusters.map(cluster => {
-                    const clusterState = getClusterState(
-                      cluster.healthy,
-                      cluster.reachable,
-                      cluster.nodeCount,
-                      undefined,
-                      cluster.errorType
-                    )
-                    const stateLabel = clusterState === 'healthy' ? '' :
-                      clusterState === 'degraded' ? 'degraded' :
-                      clusterState === 'unreachable-auth' ? 'needs auth' :
-                      clusterState === 'unreachable-timeout' ? 'offline' :
-                      'offline'
-                    return (
-                      <button
-                        key={cluster.name}
-                        onClick={() => toggleClusterFilter(cluster.name)}
-                        className={`w-full px-2 py-1.5 text-xs text-left rounded transition-colors flex items-center gap-2 ${
-                          localClusterFilter.includes(cluster.name) ? 'bg-purple-500/20 text-purple-400' : 'hover:bg-secondary text-foreground'
-                        }`}
-                        title={stateLabel ? `${cluster.name} (${stateLabel})` : cluster.name}
-                      >
-                        <ClusterStatusDot state={clusterState} size="sm" />
-                        <span className="flex-1 truncate">{cluster.name}</span>
-                        {stateLabel && (
-                          <span className="text-2xs text-muted-foreground shrink-0">{stateLabel}</span>
-                        )}
-                      </button>
-                    )
-                  })}
-                </div>
-              </div>,
-              document.body
-            )}
-          </div>
+          <LLMdClusterFilter
+            containerRef={clusterFilterRef}
+            buttonRef={clusterFilterBtnRef}
+            availableClusters={availableClusters}
+            selectedClusters={localClusterFilter}
+            showDropdown={showClusterFilter}
+            dropdownStyle={dropdownStyle}
+            onToggleDropdown={() => setShowClusterFilter(!showClusterFilter)}
+            onClearFilter={() => setLocalClusterFilter([])}
+            onToggleCluster={toggleClusterFilter}
+          />
         )}
         <button
           onClick={handleRefresh}
@@ -675,72 +563,12 @@ export function LLMdStackMonitor({ config: _config }: LLMdStackMonitorProps) {
           />
 
           {/* Component sections */}
-          <div className="flex-1 overflow-y-auto space-y-0.5">
-            {sections.map(section => {
-              const SectionIcon = section.icon
-              const isExpanded = expandedSections.has(section.label)
-              const sectionHealthy = section.items.filter(i => i.status === 'healthy').length
-              const allHealthy = sectionHealthy === section.items.length
-
-              return (
-                <div key={section.label} className="border-b border-border/30 last:border-0">
-                  <button
-                    onClick={() => toggleSection(section.label)}
-                    className="w-full flex items-center gap-2 py-1.5 px-1 text-left hover:bg-card/30 rounded transition-colors"
-                  >
-                    {isExpanded
-                      ? <ChevronDown className="w-3 h-3 text-muted-foreground shrink-0" />
-                      : <ChevronRight className="w-3 h-3 text-muted-foreground shrink-0" />}
-                    <SectionIcon className={cn('w-3.5 h-3.5 shrink-0', section.color)} />
-                    <span className="text-sm text-foreground flex-1">{section.label}</span>
-                    <span
-                      className={cn(
-                        'text-xs px-1.5 py-0.5 rounded cursor-default',
-                        allHealthy ? 'bg-green-500/20 text-green-400' : 'bg-yellow-500/20 text-yellow-400',
-                      )}
-                      title={`${sectionHealthy} healthy out of ${section.items.length} total ${section.label} components`}
-                    >
-                      {sectionHealthy}/{section.items.length}
-                    </span>
-                  </button>
-                  {isExpanded && (
-                    <div className="ml-8 mb-1.5 space-y-0.5">
-                      {section.items.map((item, idx) => (
-                        <div key={`${section.label}-${idx}-${item.name}`} className="flex items-center gap-2 py-0.5 px-1 rounded hover:bg-card/30 transition-colors group">
-                          <div className={cn('w-1.5 h-1.5 rounded-full shrink-0', STATUS_DOT[item.status] || 'bg-gray-400')} />
-                          <span className="text-xs text-foreground truncate flex-1">{item.name}</span>
-                          {item.namespace && (
-                            <StatusBadge color="purple" size="xs" className="shrink-0">
-                              {item.namespace}
-                            </StatusBadge>
-                          )}
-                          {item.detail && (
-                            <span className="text-2xs text-muted-foreground shrink-0 truncate max-w-[150px]">
-                              {item.detail}
-                            </span>
-                          )}
-                          {item.cluster && (
-                            <span className="text-2xs px-1 py-0.5 rounded bg-secondary text-muted-foreground shrink-0">
-                              {item.cluster}
-                            </span>
-                          )}
-                          {/* Diag icon - show for non-healthy items */}
-                          {item.status !== 'healthy' && (
-                            <CardAIActions
-                              resource={{ kind: 'Deployment', name: item.name, namespace: item.namespace, cluster: item.cluster, status: item.status }}
-                              issues={item.detail ? [{ name: item.status, message: item.detail }] : []}
-                              showRepair={false}
-                              onDiagnose={(e) => { e.stopPropagation(); handleItemDiagnose(item) }}
-                            />
-                          )}
-                        </div>
-                      ))}
-                    </div>
-                  )}
-                </div>
-              )
-            })}
-          </div>
+          <LLMdComponentSections
+            sections={sections}
+            expandedSections={expandedSections}
+            onToggleSection={toggleSection}
+            onDiagnoseItem={handleItemDiagnose}
+          />
 
           {/* Pagination */}
           {needsPagination && (
@@ -793,65 +621,11 @@ export function LLMdStackMonitor({ config: _config }: LLMdStackMonitorProps) {
           />
 
           {/* Issues list */}
-          <div className="flex-1 overflow-y-auto space-y-2">
-            {paginatedIssues.length > 0 ? (
-              paginatedIssues.map(issue => {
-                const severityConfig = {
-                  critical: { bg: 'bg-red-500/10', border: 'border-red-500/30', text: 'text-red-400', badge: 'bg-red-500/20 text-red-400', icon: 'text-red-400' },
-                  warning: { bg: 'bg-yellow-500/10', border: 'border-yellow-500/30', text: 'text-yellow-400', badge: 'bg-yellow-500/20 text-yellow-400', icon: 'text-yellow-400' },
-                  info: { bg: 'bg-blue-500/10', border: 'border-blue-500/30', text: 'text-blue-400', badge: 'bg-blue-500/20 text-blue-400', icon: 'text-blue-400' } }
-                const config = severityConfig[issue.severity as keyof typeof severityConfig] || severityConfig.info
-
-                return (
-                  <div
-                    key={issue.id}
-                    className={cn('rounded-lg p-3 border', config.bg, config.border)}
-                  >
-                    <div className="flex items-start gap-2">
-                      <AlertTriangle className={cn('w-4 h-4 mt-0.5 shrink-0', config.icon)} />
-                      <div className="flex-1 min-w-0">
-                        <div className="flex items-center gap-2 flex-wrap">
-                          <span className={cn('text-sm font-medium', config.text)}>{issue.title}</span>
-                          <span className={cn('text-2xs px-1.5 py-0.5 rounded', config.badge)}>{issue.severity}</span>
-                        </div>
-                        {issue.description && (
-                          <p className="text-xs text-muted-foreground mt-1">{issue.description}</p>
-                        )}
-                        <div className="flex items-center gap-2 mt-2 flex-wrap">
-                          {issue.resource?.namespace && (
-                            <StatusBadge color="purple" size="xs">
-                              {issue.resource.namespace}
-                            </StatusBadge>
-                          )}
-                          {issue.resource?.cluster && (
-                            <span className="text-2xs px-1.5 py-0.5 rounded bg-secondary text-muted-foreground">
-                              {issue.resource.cluster}
-                            </span>
-                          )}
-                        </div>
-                      </div>
-                      <CardAIActions
-                        resource={{ kind: issue.resource?.kind || 'Resource', name: issue.resource?.name || issue.title, namespace: issue.resource?.namespace, cluster: issue.resource?.cluster, status: issue.severity }}
-                        issues={[{ name: issue.title, message: issue.description || '' }]}
-                        showRepair={false}
-                        onDiagnose={() => handleItemDiagnose({
-                          name: issue.resource?.name || issue.title,
-                          status: issue.severity === 'critical' ? 'unhealthy' : 'degraded',
-                          namespace: issue.resource?.namespace,
-                          cluster: issue.resource?.cluster })}
-                      />
-                    </div>
-                  </div>
-                )
-              })
-            ) : (
-              <div className="flex flex-col items-center justify-center py-8 text-center text-muted-foreground">
-                <AlertTriangle className="w-8 h-8 opacity-30 mb-2" />
-                <p className="text-sm">{issueSearch ? 'No issues match your search' : 'No issues detected'}</p>
-                {!issueSearch && <p className="text-xs opacity-70 mt-1">All components are healthy</p>}
-              </div>
-            )}
-          </div>
+          <LLMdIssuesList
+            issues={paginatedIssues}
+            searchQuery={issueSearch}
+            onDiagnoseItem={handleItemDiagnose}
+          />
 
           {/* Pagination */}
           {needsIssuePagination && (
